feat/refactor: implement rn command (move some helpers)
