+++ src/helpers/get-path-from-arg.js
@@ -1,0 +1,5 @@
+import path from "path";
+
+export const getPathFromArg = (fileName, workingDirectory) => path.isAbsolute(fileName)
+    ? fileName
+    : path.join(workingDirectory.path, fileName);
