+++ src/working-directory.js
